+++ src/group/group.dto.ts
@@ -2,11 +2,10 @@ import { PartialType, PickType } from '@nestjs/swagger';
 import { Group } from './group.schema';
 
 export class CreateGroupDto extends PickType(Group, [
+  'name',
   'members',
 ] as const) {
 }
 
-export class UpdateGroupDto extends PartialType(PickType(Group, [
-  'members',
-] as const)) {
+export class UpdateGroupDto extends PartialType(CreateGroupDto) {
 }